Screen moves with space+leftmousekey

+++ DearImguiLearning/GlobalSettings.h
@@ -5,7 +5,7 @@ struct CameraSettings {
 	int zoomSpeed;
 	int zoomMax;
 	int zoomMin;
-	int scrollSpeed;
+	float scrollSpeed;
 };
